Move get_ssl_verify and get_ssl_certificate in a different place.
