+++ litellm/utils.py
@@ -11049,25 +11049,6 @@ def _add_key_name_and_team_to_alert(request_info: str, metadata: dict) -> str:
         return request_info
 
 
-def get_ssl_verify() -> Union[str, bool]:
-    ssl_verify_env = os.getenv("SSL_VERIFY")
-
-    if ssl_verify_env is None:
-        return litellm.ssl_verify
-
-    if isinstance(ssl_verify_env, str):
-        if ssl_verify_env.lower() == "true":
-            return True
-        elif ssl_verify_env.lower() == "false":
-            return False
-        else:
-            return ssl_verify_env
-
-
-def get_ssl_certificate() -> Optional[str]:
-    return os.getenv("SSL_CERTIFICATE", litellm.ssl_certificate)
-
-
 class ModelResponseIterator:
     def __init__(self, model_response: ModelResponse, convert_to_delta: bool = False):
         if convert_to_delta == True:
